# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "TD AngularJS-Laravel",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate an user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": "",
                        "type": "string",
                        "default": "\n    {\n    \"email\":\"ignacio@example.com\",\n    \"password\":\"test4echo\"\n    }"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
...
}